Change plural authors resource to singular and update references
The tutorial starts us out using authors as a plural resource, so that we can see all the authors in a listing and work with authors as a collection. For my purposes, users(authors) don't need to see a list of each other. I can see that being an admin feature, but at this time it is uncessary.